refactor: move margin-auto button to SettingsPanel action
Move the center-block button from MarginPadding into SpacingControls
as a SettingsPanel action, and wrap spacing in a collapsible panel
consistent with SizingControls.

Made-with: Cursor
