+++ components/ui/icon.tsx
@@ -17,7 +17,7 @@ export interface IconProps extends React.SVGProps<SVGSVGElement> {
     | 'paintbrush' | 'swatch' | 'cube' | 'triangle-down' | 'triangle-right' | 'plus-circle' | 'plus-circle-solid' | 'detach' | 'email' | 'phone' | 'paperclip' | 'unlink' | 'map'
     | 'paragraph' | 'droplet' | 'expand' | 'collapse' | 'filter' | 'crosshair' | 'lightbox' | 'slider' | 'slide' | 'slides' | 'slide-button-prev' | 'slide-button-next'
     | 'slide-bullets' | 'slide-bullet' | 'slide-navigation' | 'slide-fraction' | 'loop-alternate' | 'loop-repeat' | 'listItem' | 'external-link'
-    | 'settings'
+    | 'settings' | 'center-block'
   );
 }
 
@@ -39,6 +39,11 @@ const ICONS: Record<IconProps['name'], React.ReactNode> = {
   layers: (
     <>
       <path d="M11.1025725,7.91284093 L11.7212706,8.235504 C11.893462,8.32524748 11.9992501,8.48939425 11.9992501,8.66683254 C11.9992501,8.84427082 11.893462,9.0084176 11.7212706,9.09816107 L6.28422618,11.9314629 C6.10868449,12.0228457 5.89131551,12.0228457 5.71577382,11.9314629 L0.278729391,9.09816107 C0.106538049,9.0084176 0.000749937162,8.84427082 0.000749937162,8.66683254 C0.000749937162,8.48939425 0.106538049,8.32524748 0.278729391,8.235504 L0.89742755,7.91350759 L5.14657151,10.127483 C5.67357437,10.4021523 6.32642563,10.4021523 6.85342849,10.127483 L11.1025725,7.91284093 Z M11.1025725,5.24620394 L11.7212706,5.56886701 C11.893462,5.65861049 11.9992501,5.82275726 11.9992501,6.00019555 C11.9992501,6.17763383 11.893462,6.3417806 11.7212706,6.43152408 L6.28422618,9.26482588 C6.10868449,9.35620872 5.89131551,9.35620872 5.71577382,9.26482588 L0.278729391,6.43152408 C0.106538049,6.3417806 0.000749937162,6.17763383 0.000749937162,6.00019555 C0.000749937162,5.82275726 0.106538049,5.65861049 0.278729391,5.56886701 L0.89742755,5.2468706 L5.14657151,7.46084596 C5.67357437,7.73551529 6.32642563,7.73551529 6.85342849,7.46084596 L11.1025725,5.24620394 Z M5.71502388,0.0689282154 C5.89094307,-0.0229760718 6.10905693,-0.0229760718 6.28497612,0.0689282154 L11.7220205,2.90223002 C11.8942119,2.99197349 12,3.15612027 12,3.33355855 C12,3.51099684 11.8942119,3.67514361 11.7220205,3.76488709 L6.28497612,6.59818889 C6.10905693,6.69009318 5.89094307,6.69009318 5.71502388,6.59818889 L0.277979454,3.76488709 C0.105788111,3.67514361 0,3.51099684 0,3.33355855 C0,3.15612027 0.105788111,2.99197349 0.277979454,2.90223002 Z" />
+    </>
+  ),
+  'center-block': (
+    <>
+      <path d="M1 9v2h10V9h1v2.25a.75.75 0 0 1-.75.75H.75a.75.75 0 0 1-.75-.75V9h1Zm7.5-7a.5.5 0 0 1 .5.5v7a.5.5 0 0 1-.5.5h-5a.5.5 0 0 1-.5-.5v-7a.5.5 0 0 1 .5-.5h5ZM8 3H4v6h4V3ZM.784 4.089l.07.057 1.5 1.5a.5.5 0 0 1 .057.638l-.057.07-1.5 1.5a.5.5 0 0 1-.765-.638l.057-.07L1.293 6 .146 4.854a.5.5 0 0 1-.057-.638l.057-.07a.5.5 0 0 1 .638-.057Zm11.07.057a.5.5 0 0 1 .057.638l-.057.07L10.707 6l1.147 1.146a.5.5 0 0 1 .057.638l-.057.07a.5.5 0 0 1-.638.057l-.07-.057-1.5-1.5a.5.5 0 0 1-.057-.638l.057-.07 1.5-1.5a.5.5 0 0 1 .708 0ZM11.25 0a.75.75 0 0 1 .75.75V3h-1V1H1v2H0V.75A.75.75 0 0 1 .75 0h10.5Z" />
     </>
   ),
   eyedrop: (
